refactory the code to remove redundant code in the systems
