--- cmd/factory/compose/api_server_test.go
@@ -1,138 +1,0 @@
-package compose_test
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"io"
-	"net"
-	"net/http"
-	"os"
-	"path/filepath"
-	"testing"
-	"time"
-
-	"github.com/portpowered/infinite-you/cmd/factory/compose"
-	"github.com/portpowered/infinite-you/pkg/config"
-	"github.com/portpowered/infinite-you/pkg/interfaces"
-	"github.com/portpowered/infinite-you/pkg/service"
-	"github.com/portpowered/infinite-you/pkg/testutil/factoryfixtures"
-	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
-	"github.com/portpowered/infinite-you/pkg/transports/mapping"
-	"go.uber.org/zap"
-)
-
-func TestServeAPIServer_WiredFactoryServiceServesStatus(t *testing.T) {
-	t.Parallel()
-
-	dir := t.TempDir()
-	factoryfixtures.WriteFactoryJSON(t, dir, factoryfixtures.MinimalFactoryConfig())
-	writeComposeTestWorkerAgentsMD(t, dir, "worker-a")
-	writeComposeTestWorkstationAgentsMD(t, dir, "process")
-	if err := os.MkdirAll(filepath.Join(dir, interfaces.InputsDir), 0o755); err != nil {
-		t.Fatalf("create inputs dir: %v", err)
-	}
-
-	listener, err := net.Listen("tcp", "127.0.0.1:0")
-	if err != nil {
-		t.Fatalf("Listen: %v", err)
-	}
-	port := listener.Addr().(*net.TCPAddr).Port
-	if err := listener.Close(); err != nil {
-		t.Fatalf("Close listener: %v", err)
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	ready := make(chan struct{})
-	svcCfg := &service.FactoryServiceConfig{
-		Dir:               dir,
-		RuntimeMode:       interfaces.RuntimeModeService,
-		MockWorkersConfig: config.NewEmptyMockWorkersConfig(),
-		Port:              port,
-		Logger:            zap.NewNop(),
-		APIServerStarter: func(ctx context.Context, runtime apisurface.APISurface, bindPort int, l *zap.Logger) error {
-			apiListener, listenErr := net.Listen("tcp", fmt.Sprintf("127.0.0.1:%d", bindPort))
-			if listenErr != nil {
-				return listenErr
-			}
-			close(ready)
-			return compose.ServeAPIServer(ctx, runtime, bindPort, l, apiListener)
-		},
-	}
-
-	svc, err := compose.InjectFactoryService(ctx, svcCfg)
-	if err != nil {
-		t.Fatalf("InjectFactoryService: %v", err)
-	}
-
-	runErrCh := make(chan error, 1)
-	go func() {
-		runErrCh <- svc.Run(ctx)
-	}()
-
-	select {
-	case <-ready:
-	case err := <-runErrCh:
-		t.Fatalf("Run exited before API server ready: %v", err)
-	case <-time.After(10 * time.Second):
-		t.Fatal("timed out waiting for API server starter")
-	}
-
-	baseURL := fmt.Sprintf("http://127.0.0.1:%d", port)
-	var status factoryapi.StatusResponse
-	deadline := time.Now().Add(10 * time.Second)
-	for time.Now().Before(deadline) {
-		resp, err := http.Get(baseURL + "/status")
-		if err != nil {
-			time.Sleep(50 * time.Millisecond)
-			continue
-		}
-		body, readErr := io.ReadAll(resp.Body)
-		resp.Body.Close()
-		if readErr != nil {
-			t.Fatalf("read /status body: %v", readErr)
-		}
-		if resp.StatusCode != http.StatusOK {
-			time.Sleep(50 * time.Millisecond)
-			continue
-		}
-		if err := json.Unmarshal(body, &status); err != nil {
-			t.Fatalf("decode /status: %v", err)
-		}
-		break
-	}
-	if status.FactoryState == "" {
-		t.Fatalf("GET /status factory_state empty after polling %s/status", baseURL)
-	}
-
-	cancel()
-	if err := <-runErrCh; err != nil && err != context.Canceled {
-		t.Fatalf("Run: %v", err)
-	}
-}
-
-func writeComposeTestWorkerAgentsMD(t *testing.T, factoryDir, workerName string) {
-	t.Helper()
-	workerDir := filepath.Join(factoryDir, "workers", workerName)
-	if err := os.MkdirAll(workerDir, 0o755); err != nil {
-		t.Fatalf("create worker dir: %v", err)
-	}
-	content := "---\ntype: MODEL_WORKER\nmodel: claude-3-5-haiku-20241022\n---\nYou are a helpful assistant.\n"
-	if err := os.WriteFile(filepath.Join(workerDir, "AGENTS.md"), []byte(content), 0o644); err != nil {
-		t.Fatalf("write worker AGENTS.md: %v", err)
-	}
-}
-
-func writeComposeTestWorkstationAgentsMD(t *testing.T, factoryDir, workstationName string) {
-	t.Helper()
-	workstationDir := filepath.Join(factoryDir, "workstations", workstationName)
-	if err := os.MkdirAll(workstationDir, 0o755); err != nil {
-		t.Fatalf("create workstation dir: %v", err)
-	}
-	content := "---\ntype: MODEL_WORKSTATION\n---\nDo the work.\n"
-	if err := os.WriteFile(filepath.Join(workstationDir, "AGENTS.md"), []byte(content), 0o644); err != nil {
-		t.Fatalf("write workstation AGENTS.md: %v", err)
-	}
-}
