--- cmd/factory/compose/api_transport.go
@@ -1,13 +1,0 @@
-package compose
-
-import (
-	"context"
-
-	"github.com/portpowered/infinite-you/pkg/initializer"
-)
-
-// InjectAPITransport composes API handler dependencies through pkg/initializer
-// without using the wire FactoryService composition root.
-func InjectAPITransport(ctx context.Context, cfg *initializer.Config) (*initializer.APITransport, error) {
-	return initializer.InitializeAPITransport(ctx, cfg)
-}
