--- cmd/factory/compose/generate.go
@@ -1,3 +1,0 @@
-package compose
-
-//go:generate go run -mod=mod github.com/google/wire/cmd/wire
