+++ pkg/service/factory_build.go
@@ -38,7 +38,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/service/runtimebuild"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/dashboardrender"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
-	"github.com/portpowered/infinite-you/pkg/transports/mapping"
+	apisurface "github.com/portpowered/infinite-you/pkg/transports/mapping"
 	"github.com/portpowered/infinite-you/pkg/workers"
 	workerexecutor "github.com/portpowered/infinite-you/pkg/workers/executor"
 	workeragentrun "github.com/portpowered/infinite-you/pkg/workers/executor/agentrun"
@@ -240,7 +240,6 @@ const (
 	runtimeMetricStateIdle            = "runtime.state.idle"
 	runtimeMetricStatePaused          = "runtime.state.paused"
 	runtimeMetricStateFailed          = "runtime.state.failed"
-	runtimeMetricQueueInFlight        = "runtime.queue.in_flight"
 	runtimeMetricQueueSubmissionCount = "queue.submission_count"
 	runtimeMetricDispatchStarted      = "dispatch.started"
 	runtimeMetricDispatchComplete     = "dispatch.completed"
