+++ pkg/wire/generate.go
@@ -1,0 +1,3 @@
+package wire
+
+//go:generate go run github.com/google/wire/cmd/wire
