--- tasks/todo/factory-session-load-failure-debuggability-prd.json
@@ -1,87 +1,0 @@
-{
-  "project": "Factory Session Load Failure Debuggability",
-  "branchName": "factory-session-load-failure-debuggability-prd",
-  "description": "Improve debuggability when POST /factory-sessions inspects a readable folder but cannot recognize a runnable factory because runtime config loading fails during discovery. Preserve the real failure in logs and API classification, and show actionable inline dashboard diagnostics instead of suggesting factory initialization.",
-  "context": {
-    "customerAsk": "Improve debuggability for factory session target discovery so readable folders with broken factory configs do not collapse into not_runnable or init-new-factory behavior. Log the real load failure, return a structured validation-style API error, and update the dashboard to show actionable issues instead of suggesting factory initialization.",
-    "problem": "Today a runtime config load failure during POST /factory-sessions discovery can be silently suppressed into not_runnable or validate-only initsNewFactory: true. That hides the real failure from operators, prevents API consumers from distinguishing broken configs from empty folders, and can mislead dashboard users into initializing a new factory over an existing broken one.",
-    "solution": "Preserve config-load failures as a first-class discovery outcome in the backend, emit structured operator logs, shape a safe validation-style BAD_REQUEST API response with a stable failure code and issue targets, and update the dashboard launch flow to render accessible inline diagnostics while keeping empty-folder initialization behavior unchanged."
-  },
-  "acceptanceCriteria": [
-    "POST /factory-sessions preserves readable-folder runtime config load failures as a distinct classification rather than collapsing them into not_runnable or initNewFactory behavior.",
-    "The backend emits structured error logs for discovery-time config load failures with submitted folder path, probed target path, and target identity when known.",
-    "Validate-only and session-open discovery requests classify the same broken folder the same way while preserving existing runnable-factory and genuine empty-folder behavior.",
-    "The public API returns a structured BAD_REQUEST response with a stable config-load-failed code, safe summary message, and issue targets for the affected factory subject.",
-    "The dashboard renders an inline accessible error panel for config-load-failed responses and does not offer init-new-factory actions or create a failed session tab for that case.",
-    "Typecheck, lint, generated-artifact checks, and targeted backend, API, and UI tests pass."
-  ],
-  "userStories": [
-    {
-      "id": "factory-session-load-failure-debuggability-prd-001",
-      "title": "Log config load failures during target discovery",
-      "description": "As an operator, I want discovery-time config load failures recorded in structured logs so that I can tell why a readable folder was rejected without reproducing under a debugger.",
-      "acceptanceCriteria": [
-        "When /factory-sessions discovery probes a readable candidate and runtime config loading fails, the backend emits a structured error log instead of silently suppressing the failure.",
-        "The log includes the submitted folder path, the probed factory target path, target identity when known, and a safe summary of the underlying load failure.",
-        "The log is emitted for both validate-only and session-open discovery paths and is not emitted for successful discovery.",
-        "Typecheck passes",
-        "Tests pass"
-      ],
-      "priority": 1,
-      "passes": true,
-      "notes": ""
-    },
-    {
-      "id": "factory-session-load-failure-debuggability-prd-002",
-      "title": "Return structured API diagnostics for broken factory config loads",
-      "description": "As an API consumer, I want a structured validation-style error for broken factory config loads so that I can distinguish them from missing folders, unreadable folders, and genuine init-new-factory cases.",
-      "acceptanceCriteria": [
-        "When a readable folder exists but runtime config loading fails during /factory-sessions discovery, the API returns BAD_REQUEST rather than initsNewFactory: true or a generic not_runnable outcome.",
-        "The response includes a stable public failure code that is distinct from missing, unreadable, non-directory, and not-runnable classifications.",
-        "The response includes a safe top-level message and structured issue targets that identify the affected factory subject and summarize the load failure without exposing stack traces or unsafe internals.",
-        "The same broken folder yields the same failure classification and issue payload shape for validate-only and session-open requests.",
-        "Validate-only requests for genuinely empty readable folders still return init-new-factory metadata rather than the new config-load-failed error.",
-        "If the public API shape changes, the authored OpenAPI fragments, bundled contract, and generated Go and TypeScript clients are updated together.",
-        "Typecheck passes",
-        "Tests pass"
-      ],
-      "priority": 2,
-      "passes": true,
-      "notes": ""
-    },
-    {
-      "id": "factory-session-load-failure-debuggability-prd-003",
-      "title": "Show actionable inline diagnostics in the dashboard",
-      "description": "As a dashboard user, I want broken-factory launch failures shown inline with actionable details so that I understand what needs fixing without being sent down the new-factory path.",
-      "acceptanceCriteria": [
-        "When the dashboard receives the new config-load-failed response from /factory-sessions, it renders an inline error panel on the current launch flow instead of treating the folder as an init-new-factory candidate.",
-        "The panel shows the top-level message and the structured issue list returned by the API, with readable text and semantics that do not rely on color alone.",
-        "The UI does not create, switch to, or partially initialize a new session tab for the failed request.",
-        "Existing init-new-factory affordances remain available only for genuine empty-folder validation responses.",
-        "The failure presentation works on desktop and mobile widths and preserves keyboard-readable access to the returned issues.",
-        "Typecheck passes",
-        "Tests pass",
-        "Verify in browser using dev-browser skill"
-      ],
-      "priority": 3,
-      "passes": true,
-      "notes": ""
-    },
-    {
-      "id": "factory-session-load-failure-debuggability-prd-004",
-      "title": "Keep validate and open-session flows aligned across runnable, empty, and broken folders",
-      "description": "As a maintainer, I want validation and open-session discovery to classify the same folder the same way so that backend, API, and UI behavior do not diverge across preflight and launch paths.",
-      "acceptanceCriteria": [
-        "The same broken factory folder yields the same config-load-failed classification whether the request is validate-only or a direct session-open attempt.",
-        "A runnable factory folder remains discoverable and openable in both validate-only and session-open flows without the new error classification.",
-        "A truly empty readable folder still returns init-new-factory discovery metadata only in validate-only flows and does not regress into a broken-config classification.",
-        "Regression coverage proves the distinction between runnable-factory, empty-folder, and broken-config cases at the service/API layer and for the dashboard launch behavior that depends on those outcomes.",
-        "Typecheck passes",
-        "Tests pass"
-      ],
-      "priority": 4,
-      "passes": true,
-      "notes": ""
-    }
-  ]
-}
