+++ tests/functional/internal/support/api_server.go
@@ -8,7 +8,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/portpowered/infinite-you/cmd/factory/compose"
+	"github.com/portpowered/infinite-you/pkg/compose"
 	"github.com/portpowered/infinite-you/pkg/config"
 	"github.com/portpowered/infinite-you/pkg/factory"
 	"github.com/portpowered/infinite-you/pkg/factory/state"
@@ -18,7 +18,7 @@ import (
 	"github.com/portpowered/infinite-you/pkg/testutil/testdeps"
 	api "github.com/portpowered/infinite-you/pkg/transports/http"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
-	"github.com/portpowered/infinite-you/pkg/transports/mapping"
+	apisurface "github.com/portpowered/infinite-you/pkg/transports/mapping"
 	"go.uber.org/zap"
 )
