+++ tests/functional/internal/support/cmd/browser_api_harness/main.go
@@ -13,13 +13,13 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/portpowered/infinite-you/cmd/factory/compose"
+	"github.com/portpowered/infinite-you/pkg/compose"
 	"github.com/portpowered/infinite-you/pkg/config"
 	workflowsource "github.com/portpowered/infinite-you/pkg/orchestrators/javascript/source"
 	"github.com/portpowered/infinite-you/pkg/service"
 	api "github.com/portpowered/infinite-you/pkg/transports/http"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
-	"github.com/portpowered/infinite-you/pkg/transports/mapping"
+	apisurface "github.com/portpowered/infinite-you/pkg/transports/mapping"
 	"go.uber.org/zap"
 )
